api/requirements.txt
# Supabase Python client
supabase>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
